fixed deleting previous refresh tokens

+++ backend/src/main/java/com/smartship/backend/app/utility/RefreshTokenUtil.java
@@ -33,7 +33,9 @@ public class RefreshTokenUtil {
      */
     public RefreshToken createRefreshToken(Long userId) {
         // First delete if user already has a refreshToken
-        refreshTokenRepository.deleteAllByUserId(userId);
+        if (refreshTokenRepository.existsByUserId(userId)) {
+            refreshTokenRepository.deleteAllByUserId(userId);
+        }
 
         // Give new token
         RefreshToken refreshToken = new RefreshToken();
